refactor: streamline imports and update type usage across action files
